+++ src/Util/EnvironmentInfoProvider.php
@@ -53,7 +53,7 @@ class EnvironmentInfoProvider
     /**
      * @var string plugin version
      */
-    public const PLUGIN_VERSION = '1.1.4';
+    public const PLUGIN_VERSION = '1.2.0';
 
     /**
      * @var string Name of Lengow front controller